Convert program to draw a sphere
Replace the vertex and fragment shaders with those from Program 6.1 - Sphere.

Convert init and display methods to draw a sphere instead of a pyramid and cube.

Add the earth texture for temporary testing along with the loadTexture method.

Implement the KeyListener interface in preparation for controlling the camera with keyboard input.