Add rotation to sun
Add an FPSAnimator to repeatedly call the display method.

In display, rotate the sun's model matrix around the y-axis by a value dependent on the system elapsed time.

Change the texture from the earth to the sun.

Move the sun to the origin (0, 0, 0).

Move the camera back along the z-axis to achieve a wider field of view (0, 0, 10).