Add orbiting mechanics and moon for the earth
Make the earth orbit around the sun on the x-z plane using the sin and cos functions. The argument to those functions is based on the system elapsed time.

Scale the earth to 0.75x base size.

Enable depth testing so the earth will actually appear to go behind the sun.

Add a moon for the earth. The moon rotates on its z-axis and orbits around the earth on the earth's y-z plane. It is scaled to 0.25x base size.