Add Phobos and streamline vertex setup
Create a new method setupSphereVertices, which takes a Sphere and the VBO index to start from. It performs all operations previously contained within the setupVertices method, but with a modular setup to reduce code duplication. Unfortunately, the same cannot be done for the code in the display method.

Clean up member variables. Remove superfluous ones and move related ones of the same type to the same line to save space.

Add Phobos, Mars's moon. It orbits around Mars on Mars's (y = -x) plane at twice the normal speed. It rotates on its y-z plane and is 0.2x the base sphere size.