Attempt to implement some camera control
Store the view matrix as a member variable. That way, we can manipulate it when certain keys are pressed.

Implement strafing. This works as long as the camera is not rotated with the arrow keys. I am still working on how to get the strafe axes to conform to the camera's UVN axes rather than the XYZ axes.