Fix literally everything about camera control
Store camera pitch and yaw in member variables. When rotating the camera, add/subtract values from these.

Store the forward vector (pointing in the direction the camera is looking) in a member variable. When strafing, use the forward vector to calculate side and top vectors (if needed) and add components to the XYZ coordinates of the camera.

Update forward vector every time display is called using sine and cosine transformations based on camera pitch and yaw.

Use call to lookAt method at view matrix. Pass camera position as the camera position (obviously), the forward vector relative to the camera position as the target point, and the y-axis as the up vector.

Draw world axes in the scene and make them toggleable with Space.

Make some progress with the pentagonal prism class.